IMP stock ean 128
