--- select_lot_by_ean128/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-import stock
