--- select_lot_by_ean128/__openerp__.py
@@ -1,30 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    'name': 'Select Lot By EAN128',
-    'version': '1.0',
-    'category': 'Warehouse Management',
-    'sequence': 14,
-    'summary': '',
-    'description': """
-Select Lot By EAN128
-====================
-
-    """,
-    'author':  'ADHOC',
-    'website': 'www.adhoc.com.ar',
-    'images': [
-    ],
-    'depends': [
-        'stock_ean128',
-    ],
-    'data': [
-             ],
-    'demo': [
-    ],
-    'test': [
-    ],
-    'installable': True,
-    'auto_install': False,
-    'application': False,
-}
-# vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
