+++ stock_ean128/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 import stock
+import product
