+++ stock_ean128/stock.py
@@ -2,7 +2,7 @@
 from openerp import models, fields, api
 
 
-class stock_picking(models.Model):
+class stock_production_lot(models.Model):
     _inherit = 'stock.production.lot'
 
     ean_128 = fields.Char(string="EAN128")
@@ -19,3 +19,12 @@ class stock_picking(models.Model):
         else:
             name += '(17)' + 'N/A'
         self.ean_128 = name
+
+    def name_search(
+            self, cr, uid, name, args=None, operator='ilike', context=None, limit=100):
+        res = super(stock_production_lot, self).name_search(
+            cr, uid, name=name, args=args, operator=operator, context=context, limit=limit)
+        ids = self.search(cr, uid, [('ean_128', operator, name)],
+                          limit=limit, context=context)
+        res += self.name_get(cr, uid, ids, context=context)
+        return res
